+++ packages/debian/deb.sh
@@ -3,11 +3,13 @@
 export RUST_MIN_STACK=67108864
 
 cargo run --release --bin hnefatafl-ai -- --man --username ""
+cargo run --release --bin hnefatafl-server -- --man
 cargo run --release --bin hnefatafl-text-protocol -- --man
 cargo run --release --example hnefatafl-client -- --man
 cargo run --release -- --man
 
 gzip --no-name --best hnefatafl-ai.1
+gzip --no-name --best hnefatafl-server.1
 gzip --no-name --best hnefatafl-text-protocol.1
 gzip --no-name --best hnefatafl-client.1
 gzip --no-name --best hnefatafl-server-full.1
@@ -15,12 +17,13 @@ gzip --no-name --best hnefatafl-server-full.1
 PACKAGE=$(cargo deb)
 
 rm hnefatafl-ai.1.gz
+rm hnefatafl-server.1.gz
 rm hnefatafl-text-protocol.1.gz
 rm hnefatafl-client.1.gz
 rm hnefatafl-server-full.1.gz
 
 echo $PACKAGE
-lintian $PACKAGE
+lintian -EviIL +pedantic $PACKAGE
 
 if [ -z $1 ]; then
     exit
@@ -29,6 +32,5 @@ fi
 if [ $1 = 'install' ]; then
     sudo dpkg --remove hnefatafl-copenhagen
     sudo dpkg --install $PACKAGE
-    sudo systemctl restart hnefatafl.service
     sudo systemctl daemon-reload
 fi